Dictate certdir to sign_domain by command_sign_domains
Now the first parameter is the certificate output directory. This is
dictated by command_sign_domains as a single point of authority.
Also in sign_domain and command_sign_domains, the variables have been
rephrased to idicate that the first domain is only the primary, which
is put into the CN field, being part of alldomains. The primary still
indicates the configuration file in DOMAINS_D and the output directory.