Move hook challenge deployment into loop
Certain DNS services (Duck DNS) only allow a single TXT record for a
domain and all subdomains.  So the last TXT value will overwrite all
early values.  To support such a service, setting TXT records
must be done in the challenge loop.  That way the TXT record is set and
immediately checked before going on to the next one.