fix hook based challenges if the same domain appears twice or more
When two or more domains are present in the challenge, all the hook
calls happen before asking the acme server to verify the challenge.
This is breaking configurations that need to verify the same domain
more than once, e.g. dns based challenges with `example.com`. and
`*.example.com`. In this case, `example.com` will be verified twice,
maybe with distinct tokens, leading to the second configuration
overwriting the first one before the challenge verification.

This update moves the hook call into the while-loop that asks the acme
server to verify the challenge, so a second configuration will only
happen after the first one has being verified.